+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+2025-07-18 - [0.2]:
+- Adding cache expiration
+- Fixing logs
+- Add a message when a user first gets into contact with the bot.
+- Add a list of allowed users
+- Find which model is cheapest and create key with limitation
+
 2025-07-17 - [0.1]:
 - Initial version 0.1
 
 BUGS:
-- For some reasons the logs are not correctly written.
+
+TODO: